The server can now ping clients to make sure they are still connected

+++ LambdaEngine/Include/Time/API/Timestamp.h
@@ -21,27 +21,27 @@ namespace LambdaEngine
 		{
 		}
 
-		FORCEINLINE float AsSeconds() const
+		FORCEINLINE float64 AsSeconds() const
 		{
-			constexpr float SECONDS = 1000.0f * 1000.0f * 1000.0f;
-			return float(m_NanoSeconds) / SECONDS;
+			constexpr float64 SECONDS = 1000.0 * 1000.0 * 1000.0;
+			return float64(m_NanoSeconds) / SECONDS;
 		}
 
-		FORCEINLINE float AsMilliSeconds() const
+		FORCEINLINE float64 AsMilliSeconds() const
 		{
-			constexpr float MILLISECONDS = 1000.0f * 1000.0f;
-			return float(m_NanoSeconds) / MILLISECONDS;
+			constexpr float64 MILLISECONDS = 1000.0 * 1000.0;
+			return float64(m_NanoSeconds) / MILLISECONDS;
 		}
 
-		FORCEINLINE float AsMicroSeconds() const
+		FORCEINLINE float64 AsMicroSeconds() const
 		{
-			constexpr float MICROSECONDS = 1000.0f;
-			return float(m_NanoSeconds) / MICROSECONDS;
+			constexpr float64 MICROSECONDS = 1000.0;
+			return float64(m_NanoSeconds) / MICROSECONDS;
 		}
 
 		FORCEINLINE uint64 AsNanoSeconds() const
 		{
-			return uint64(m_NanoSeconds);
+			return m_NanoSeconds;
 		}
 
 		FORCEINLINE Timestamp& operator=(const Timestamp& other)
@@ -84,27 +84,27 @@ namespace LambdaEngine
 			return *this;
 		}
 
-		FORCEINLINE static Timestamp Seconds(float s)
+		FORCEINLINE static Timestamp Seconds(float64 s)
 		{
-			constexpr float SECOND = 1000.0f * 1000.0f * 1000.0f;
+			constexpr float64 SECOND = 1000.0 * 1000.0 * 1000.0;
 			return Timestamp(uint64(s * SECOND));
 		}
 
-		FORCEINLINE static Timestamp MilliSeconds(float ms)
+		FORCEINLINE static Timestamp MilliSeconds(float64 ms)
 		{
-			constexpr float MILLISECOND = 1000.0f * 1000.0f;
+			constexpr float64 MILLISECOND = 1000.0 * 1000.0;
 			return Timestamp(uint64(ms * MILLISECOND));
 		}
 
-		FORCEINLINE static Timestamp MicroSeconds(float us)
+		FORCEINLINE static Timestamp MicroSeconds(float64 us)
 		{
-			constexpr float MICROSECOND = 1000.0f;
+			constexpr float64 MICROSECOND = 1000.0;
 			return Timestamp(uint64(us * MICROSECOND));
 		}
 
 		FORCEINLINE static Timestamp NanoSeconds(uint64 ns)
 		{
-			return Timestamp(uint64(ns));
+			return Timestamp(ns);
 		}
 
 		friend Timestamp operator+(const Timestamp& left, const Timestamp& right);
